Make subtraction a special case of addition.
There are arguments to be made for and against this. On one hand, it is
arguably mathematically more fundamental to define a-b = a+(-b), however
we can't hope to make everything the most fundamental and have in
other places favoured direct correspondence to the visual representation
that's actually used in the code and in LaTeX renderings.

But addition and multiplication are clearly rather more primitive, and
merit enabling particularly good handling especially in the pattern-replacement
functions, and these are somewhat involved algorithms that could easily
harbour bugs. Thus it makes sense in this case to shift a bit more
load into the rendering side here.

Currently, we don't sugar the subtractions in rendering, this is next up.